[KYUUBI #2012] rename EventLoggingService and creates an correct directory
